Rave: fix real compiler warnings, silence deprecation noise
Reviewed all 304 warnings from the latest Ubuntu 24 build log. All of
them come from Rave's vendored code (none from FASER's own sources).
Broken down:

- 271 warnings (89%) are std::auto_ptr/std::unary_function deprecation
  notices - both still fully functional in C++11, just deprecated in
  later standards. Not worth patching dozens of vendored files for;
  silenced with -Wno-deprecated-declarations on Rave's own build
  instead (cmake/Externals.cmake).
- ~17 are autoconf "obsolete macro" notices from Rave's own
  configure.ac/m4 files (AC_HELP_STRING, AC_PROG_LIBTOOL, etc.) -
  cosmetic only, left alone.
- 3 were genuine compiler warnings worth fixing directly:

  - ZvresVertexReconstructor.cc:132 and Track.cc:29: -Wint-to-pointer-cast,
    both `(void*) some_int` on LP64 (int is 4 bytes, void* is 8) -
    widened through `long` first, which is exactly pointer-sized on
    every platform this project builds on (Linux/macOS, all LP64).
    Zero behavior change, just makes the existing intent (stash an int
    ID inside an opaque void* "private data" slot) explicit to the
    compiler.

  - KinematicParticle.cc/.h: p4() returned `const rave::Vector4D&`, but
    its body returns Vector7D::p4()'s result - which is a Vector4D
    returned *by value*. That bound a reference to a temporary that
    was already destroyed once the caller received it - a genuine
    dangling-reference bug (-Wreturn-local-addr), not just a style
    nag. Fixed by returning Vector4D by value instead, matching what
    it actually is; no internal caller in Rave or FASER relies on
    reference identity here (grepped both trees), and a caller binding
    `const Vector4D&` to the by-value result still works correctly via
    ordinary temporary lifetime extension.

+++ rave/src/RecoVertex/ZvresReco/src/ZvresVertexReconstructor.cc
@@ -129,7 +129,7 @@ to_lcfi(const reco::TransientTrack &track, vertex_lcfi::Event *const lcfi_event)
 			   track.charge(),
 			   lcfi_cov,
 			   std::vector<int>(), // SubdetectorHitNumbers
-			   (void*) track.id() ); // private data
+			   (void*)(long) track.id() ); // private data - int widened via long (LP64) to silence -Wint-to-pointer-cast
 
   MemoryManager<vertex_lcfi::Track>::Event()->registerObject(MyTrack);
